services/voice-pipeline/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
httpx==0.28.1
python-multipart==0.0.20
numpy==2.2.1
edge-tts==6.1.18
webrtcvad-wheels==2.0.14
silero-vad>=5.1.0
openwakeword>=0.6.0
onnxruntime>=1.20.1
torch>=2.0.0
speechbrain>=1.0.1
librosa>=0.10.2
scikit-learn>=1.5.0
